add record manager code to record manager file and remove from gui file (misnamed file)
