+++ src/data/record_manager.py
@@ -1,0 +1,90 @@
+import os
+import json
+import pickle
+from typing import List, Dict, Any, Optional, Literal
+
+class RecordManager:
+    """Manage records for client, flights and airline companies. Handles CRUD operations and File Persistence."""
+    
+    RECORD_TYPES = ['client', 'flight', 'airline']
+    
+    def __init__(self, data_folder: str = "data", file_format: str = "jsonl"):
+        """Initialize RecordManager with data folder and file format."""
+        
+        self.data_folder = data_folder
+        self.file_format = file_format.lower()
+        
+        # Check if file format is supported
+        if self.file_format not in ['jsonl', 'json' 'pickle']:
+            raise ValueError(f"File format '{self.file_format}' is not supported.")
+        
+        # Create data folder if it does not exist
+        os.makedirs(self.data_folder, exist_ok=True)
+        
+        # Initialize records dictionary
+        self.records = {
+            "client": [],
+            "flight": [],
+            "airline": []
+        }
+        
+        # Load records from files
+        self.load_records()
+        
+    def _get_file_path(self, record_type: str):
+        """Get file path for record type."""
+        extention = self.file_format if self.file_format != 'jsonl' else 'json'
+        # Return file path as formatted string using record type for file extension.
+        return os.path.join(self.data_folder, f"{record_type}.{extention}")
+    
+    def load_records(self) -> None:
+        """Load all records from files."""
+        
+        for record_type in self.records.keys():
+            if record_type:
+                file_path = self._get_file_path(record_type)
+                if not os.path.exists(file_path):
+                    continue
+            
+            try:
+                if self.file_format == 'jsonl':
+                    self.records[record_type] = []
+                    with open(file_path, 'r') as file:
+                        for line in file:
+                            if line.strip(): # Check if line is not empty & strip whitespace
+                                self.records[record_type].append(json.loads(line))
+                
+                elif self.file_format == 'json':
+                    with open(file_path, 'r') as file:
+                        self.records[record_type] = json.load(file)
+                        
+                elif self.file_format == 'pickle':
+                    with open(file_path, 'rb') as file:
+                        self.records[record_type] = pickle.load(file)
+            
+            except Exception as e:
+                print(f"Error loading {record_type} records: {e}")
+                self.records[record_type] = []
+                
+    def save_records(self) -> None:
+        """Save all records to files."""
+        
+        for record_type, records in self.records.items():
+            file_path = self._get_file_path(record_type)
+            
+            try:
+                if self.file_format == 'jsonl':
+                    with open(file_path, 'w') as file:
+                        for record in records:
+                            file.write(json.dumps(record) + '\n')
+                            
+                elif self.file_format == 'json':
+                    with open(file_path, 'w') as file:
+                        json.dump(records, file, indent=4)
+                
+                elif self.file_format == 'pickle':
+                    with open(file_path, 'wb') as file:
+                        pickle.dump(records, file)
+                
+            except Exception as e:
+                print(f"Error saving {record_type} records: {e}")
